requirements.txt
# Temel kütüphaneler
discord.py==2.3.2
python-dotenv==1.0.1
ptz==0.3.0  # Discord botu için komut işleme
# Veri yapıları ve işlemleri            
# API & HTTP istekleri için
aiohttp==3.9.5              
requests==2.32.3
datatime==4.8.0  # Tarih ve saat işlemleri
cohere==3.0.0  # Cohere AI entegrasyonu
# Veri analizi ve işleme    
# JSON işleme & performans
ujson==5.9.0
ptz==0.3.0  # Discord botu için komut işleme    
# Veritabanı desteği (isteğe bağlı)
asyncpg==0.29.0     # PostgreSQL
motor==3.4.0        # MongoDB (async)

# Gelişmiş zamanlama ve görevler
apscheduler==3.10.4
# Veri doğrulama ve şemalar
pydantic==2.7.3                 
# Loglama ve hata takip (isteğe bağlı)
loguru==0.7.2
sentry-sdk==1.45.0  # Hata raporlama sistemi

# Versiyon uyumluluğu için
certifi>=2024.0.0  # Güvenli sertifikalar
charset-normalizer>=3.2.0  # Karakter kodlaması desteği         